moved defect methods to defect service
